backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
ldap3==2.9.1
httpx==0.26.0
aiosmtplib==3.0.1
email-validator==2.1.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-dotenv==1.0.1
